# backend/docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Platform bimbingan belajar online",
        "title": "paham.in API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/dev/cron/evidence-cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Jalankan manual job hapus bukti kehadiran approved yang lewat masa simpan (development)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "summary": "Run evidence cleanup job",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.RunJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dev/cron/notification-cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus notifikasi yang sudah dibaca dan lebih dari 7 hari",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Run notification cleanup job",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.RunJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dev/cron/session-cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Jalankan manual job hapus sesi kedaluwarsa (development)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "summary": "Run session cleanup job",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.RunJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dev/cron/temp-image-cleanup": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Menghapus gambar di public/temp_* yang berumur lebih dari 24 jam",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Run temp image cleanup job",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.RunJobResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/admin/dev/tables": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Status fitur reset + daftar tabel aplikasi beserta jumlah row (development)",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "summary": "List dev reset tables",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.ListTablesResponse"
                        }
                    }
                }
            }
        },
        "/admin/dev/tables/{table}": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Hapus semua row di tabel yang dipilih (development). Tabel dilindungi ditolak.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dev"
                ],
                "summary": "Reset table",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Nama tabel",
                        "name": "table",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/devreset.ResetResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/devreset.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "devreset.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "devreset.ListTablesResponse": {
            "type": "object",
            "properties": {
                "enabled": {
                    "type": "boolean"
                },
                "tables": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/devreset.TableInfo"
                    }
                }
            }
        },
        "devreset.ResetResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                },
                "table": {
                    "type": "string"
                }
            }
        },
        "devreset.RunJobResponse": {
            "type": "object",
            "properties": {
                "deleted": {
                    "type": "integer"
                },
                "job": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "devreset.TableInfo": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "protected": {
                    "type": "boolean"
                },
                "rows": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}